+++ src/org/soulspace/overarch/cli.clj
@@ -5,6 +5,7 @@
             [hawk.core :as hawk]
             [org.soulspace.overarch.core :as core]
             [org.soulspace.overarch.export :as exp]
+            [org.soulspace.overarch.render :as rndr]
             ; must be loaded here for registering of the multimethods
             ; require dynamically?
             [org.soulspace.overarch.exports.json :as json]
@@ -21,14 +22,16 @@
 (def appname "overarch")
 (def description
   "Overarch CLI Exporter
-   Reads your model and view specifications and exports them
-   into the specified format.")
+   Reads your model and view specifications and renders or exports
+   into the specified formats.")
 
 (def cli-opts
   [["-m" "--model-dir DIRNAME" "Model directory" :default "models"]
-   ["-e" "--export-dir DIRNAME" "Export directory" :default "export"]
-   ["-w" "--watch" "Watch model dir for changes and trigger export" :default false]
-   ["-f" "--format FORMAT" "Export format (graphviz, json, markdown, plantuml, structurizr)" :default :plantuml :default-desc "plantuml" :parse-fn keyword]
+   ["-r" "--render-format FORMAT" "Render format (all, graphviz, markdown, plantuml)" :parse-fn keyword] ; :default :all :default-desc "all"]
+   ["-R" "--render-dir DIRNAME" "Export directory" :default "export"]
+   ["-x" "--export-format FORMAT" "Export format (json, structurizr)" :parse-fn keyword]
+   ["-X" "--export-dir DIRNAME" "Export directory" :default "export"]
+   ["-w" "--watch" "Watch model dir for changes and trigger action" :default false]
    [nil  "--model-info" "Returns infos for the loaded model" :default false] 
    [nil  "--plantuml-list-sprites" "Lists the loaded PlantUML sprites" :default false]
 ;   [nil  "--plantuml-find-sprite" "Searches the loaded PlantUML sprites for the given name"]
@@ -87,13 +90,6 @@
 ;;;
 ;;; Handler logic
 ;;;
-
-(defn update-and-export!
-  "Read models and export the data according to the given `options`."
-  [options]
-  (core/update-state! (:model-dir options))
-  (exp/export options))
-
 (defn model-info
   "Reports information about the model and views."
   [options]
@@ -123,22 +119,35 @@
    (doseq [sprite sprite-mappings]
      (println (str (:key sprite) ": " (puml/sprite-path sprite))))))
 
-(defn handle
-  "Handle the `options` and generate the requested outputs."
+(defn dispatch
+  "Dispatch on `options` to the requested actions."
   [options]
-  (core/update-state! (:model-dir options))
-  (exp/export options)
   (when (:model-info options)
     (println (model-info options)))
   (when (:plantuml-list-sprites options)
     (print-sprite-mappings))
+  (when (:render-format options)
+    (rndr/render (:render-format options) options))
+  (when (:export-format options)
+    (exp/export (:export-format options) options)))
+
+(defn update-and-dispatch!
+  "Read models and export the data according to the given `options`."
+  [options]
+  (core/update-state! (:model-dir options))
+  (dispatch options))
+
+(defn handle
+  "Handle the `options` and generate the requested outputs."
+  [options]
+  (update-and-dispatch! options)
   (when (:watch options)
     ; TODO loop recur this update-and-export! as handler
     (hawk/watch! [{:paths [(:model-dir options)]
                    :handler (fn [ctx e]
 ;                              (println "event: " e)
 ;                              (println "context: " ctx)
-                              (update-and-export! options))}])
+                              (update-and-dispatch! options))}])
     (while true
       (Thread/sleep 5000))))
 
@@ -159,16 +168,17 @@
       (handle options))))
 
 (comment
-  (update-and-export! {:model-dir "models"
-                       :format :plantuml})
+  (update-and-dispatch! {:model-dir "models"
+                         :render-format :plantuml})
   (model-info {:model-info true})
   (print-sprite-mappings)
   (-main "--debug")
-  (-main "--debug" "--format" "json")
-  (-main "--debug" "--format" "markdown")
-  (-main "--debug" "--format" "graphviz")
-  (-main "--model-dir" "models/banking" "--format" "structurizr")
+  (-main "--debug" "--render-format" "plantuml")
+  (-main "--debug" "--render-format" "markdown")
+  (-main "--debug" "--render-format" "graphviz")
+  (-main "--debug" "--export-format" "json")
+  (-main "--model-dir" "models/banking" "--export-format" "structurizr")
   (-main "--model-info")
-  (-main "--help") ; ends REPL session
   (-main "--plantuml-list-sprites")
+  (-main "--help") ; ends REPL session
   )