+++ src/org/soulspace/overarch/view.clj
@@ -257,4 +257,3 @@ When the view is rendered hierachically, additional, not directly specified elem
   ;(collect-technologies (:elements @core/state))
   (elements-in-view (core/get-view @core/state :banking/container-view))
   (technologies-in-view (core/get-view @core/state :banking/container-view)))
-